requirements.txt
customtkinter
Pillow
tkinter
json
sys
datetime
uuid
os